Trends in cooperation on an area-level
This program creates a CSV file with the number of significant cosponsorings (cross-party) for every policy area and the contribution of that area to overall (significant) cross-party cooperation. Thus, it enebales us to see which policy area "creates" trends in cosponsorship. Input data need to be of particualr form and name (see annotations).